code/demo2/requirements.txt
langchain-openai==0.2.14
langchain-community==0.3.14
langchain-core==0.3.28
langchain-text-splitters==0.3.4
langgraph==0.2.59
faiss-cpu==1.9.0.post1
tavily-python==0.5.0
python-dotenv==1.0.1
docx2txt==0.8
